+++ src/main/java/com/nequi/challenge/contexts/franchise/infrastructure/adapters/handlers/BranchOfficeHandler.java
@@ -4,7 +4,7 @@ import com.nequi.challenge.contexts.franchise.infrastructure.adapters.dto.Branch
 import com.nequi.challenge.contexts.franchise.infrastructure.adapters.dto.BranchOfficeResponseDto;
 import com.nequi.challenge.contexts.franchise.infrastructure.adapters.dto.DtoValidator;
 import com.nequi.challenge.contexts.franchise.infrastructure.mappers.BranchOfficeMapper;
-import com.nequi.challenge.contexts.franchise.infrastructure.services.BranchOfficeService;
+import com.nequi.challenge.contexts.franchise.application.services.BranchOfficeService;
 import com.nequi.challenge.contexts.shared.infrastructure.util.BuildErrorUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
